+++ es4x/src/main/java/io/reactiverse/es4x/Runtime.java
@@ -18,34 +18,95 @@ package io.reactiverse.es4x;
 import io.reactiverse.es4x.impl.graal.GraalRuntime;
 import io.reactiverse.es4x.impl.nashorn.NashornRuntime;
 import io.vertx.core.Vertx;
+import io.vertx.core.VertxOptions;
+import io.vertx.core.json.JsonObject;
 
 import java.util.Map;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.atomic.AtomicReference;
 
-public interface Runtime<T> {
+public interface Runtime<R> {
 
-  static Runtime create() {
+  static Runtime getCurrent() {
     String rtName = System.getProperty("es4x.engine");
     // rt name takes precedence in the choice
     if (rtName == null) {
       String vmName = System.getProperty("java.vm.name");
       if (vmName != null && vmName.startsWith("GraalVM")) {
-        rtName = "GraalVM";
+        rtName = "GraalJS";
+      } else {
+        rtName = "Nashorn";
       }
     }
 
-    if (rtName != null && rtName.equalsIgnoreCase("GraalVM")) {
-      // attempt to load graal loader
-      try {
-        return new GraalRuntime();
-      } catch (RuntimeException e) {
-        // Ignore...
-      }
+    if (rtName.equalsIgnoreCase("GraalJS")) {
+      System.setProperty("es4x.engine", "GraalJS");
+      return new GraalRuntime();
     }
-    // fallback (nashorn)
-    return new NashornRuntime();
+
+    if (rtName.equalsIgnoreCase("Nashorn")) {
+      System.setProperty("es4x.engine", "Nashorn");
+      return new NashornRuntime();
+    }
+
+    System.clearProperty("es4x.engine");
+    throw new RuntimeException("Unsupported runtime: " + rtName);
   }
 
+  /**
+   * return the runtime name
+   * @return runtime name.
+   */
   String name();
 
-  Vertx vertx(Object object, T JSON, Map<String, Object> arguments);
+  /**
+   * Bootstraps a Vert.x instance
+   * @param arguments arguments
+   * @return a vertx instance
+   */
+  default Vertx vertx(Map<String, Object> arguments) {
+
+    final VertxOptions options = arguments == null ? new VertxOptions() : new VertxOptions(new JsonObject(arguments));
+
+    if (options.isClustered()) {
+      final CountDownLatch latch = new CountDownLatch(1);
+
+      final AtomicReference<Throwable> err = new AtomicReference<>();
+      final AtomicReference<Vertx> holder = new AtomicReference<>();
+
+
+      Vertx.clusteredVertx(options, ar -> {
+        if (ar.failed()) {
+          err.set(ar.cause());
+          latch.countDown();
+        } else {
+          holder.set(ar.result());
+          latch.countDown();
+        }
+      });
+
+      try {
+        latch.await();
+      } catch (InterruptedException e) {
+        throw new RuntimeException(e);
+      }
+
+      if (err.get() != null) {
+        throw new RuntimeException(err.get());
+      } else {
+        return holder.get();
+      }
+    } else {
+      return Vertx.vertx(options);
+    }
+  }
+
+  Runtime<R> registerCodec(Vertx vertx);
+
+  /**
+   * Returns a module loader for the given runtime.
+   *
+   * @return loader
+   */
+  Loader<R> loader(Vertx vertx);
 }