+++ es4x/src/main/java/io/reactiverse/es4x/VerticleFactory.java
@@ -22,6 +22,8 @@ import io.vertx.core.Vertx;
 
 public class VerticleFactory implements io.vertx.core.spi.VerticleFactory {
 
+  private final Runtime runtime = Runtime.getCurrent();
+
   private Vertx vertx;
 
   @Override
@@ -35,13 +37,17 @@ public class VerticleFactory implements io.vertx.core.spi.VerticleFactory {
   }
 
   @Override
-  public Verticle createVerticle(String verticleName, ClassLoader classLoader) throws Exception {
+  public Verticle createVerticle(String verticleName, ClassLoader classLoader) {
 
     final Loader engine;
 
     synchronized (this) {
-      // create a new CommonJS loader
-      engine = Loader.create(vertx);
+      engine = runtime
+        // now that the vertx instance fully initialized  and available,
+        // register the custom JS codec for the eventbus
+        .registerCodec(vertx)
+        // create a new CommonJS loader
+        .loader(vertx);
     }
 
     return new Verticle() {
@@ -100,7 +106,15 @@ public class VerticleFactory implements io.vertx.core.spi.VerticleFactory {
 
         // this can take some time to load so it might block the event loop
         // this is usually not a issue as it is a one time operation
-        self = engine.main(fsVerticleName);
+        try {
+          engine.enter();
+          self = engine.main(fsVerticleName);
+        } catch (RuntimeException e) {
+          startFuture.fail(e);
+          return;
+        } finally {
+          engine.leave();
+        }
 
         if (self != null) {
           if (worker) {
@@ -142,7 +156,7 @@ public class VerticleFactory implements io.vertx.core.spi.VerticleFactory {
       }
 
       @Override
-      public void stop(Future<Void> stopFuture) throws Exception {
+      public void stop(Future<Void> stopFuture) {
         if (self != null) {
           if (engine.hasMember(self, "stop")) {
             try {